Speedup unzipping
Gunzipping the specs.4.8.gz file on my PC takes about 16 seconds with `Gem.gunzip`.
It's virtually instantaneous (80ms) with zcat/gunzip (i.e., it it 200
times faster).

This patch uses zcat instead of Gem.gunzip, if it is available. (The
availability check could probably be done better, but I don't know how
to do it properly.)